+++ lib/climb/climb_event.dart
@@ -1,0 +1,13 @@
+import 'package:sendrax/models/attempt.dart';
+
+abstract class ClimbEvent {}
+
+class ClearAttemptsEvent extends ClimbEvent {}
+
+class AttemptsUpdatedEvent extends ClimbEvent {
+  AttemptsUpdatedEvent(this.attempts);
+
+  final List<Attempt> attempts;
+}
+
+class ClimbErrorEvent extends ClimbEvent {}