+++ lib/location/location_bloc.dart
@@ -3,7 +3,9 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:sendrax/location/location_event.dart';
 import 'package:sendrax/location/location_state.dart';
+import 'package:sendrax/location/location_view.dart';
 import 'package:sendrax/models/climb.dart';
+import 'package:sendrax/models/climb_repo.dart';
 import 'package:sendrax/models/location.dart';
 import 'package:sendrax/models/location_repo.dart';
 import 'package:sendrax/models/user_repo.dart';
@@ -22,11 +24,9 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
     add(ClearClimbsEvent());
     final user = await UserRepo.getInstance().getCurrentUser();
     if (user != null) {
-      climbSubscription = LocationRepo.getInstance()
-          .getClimbsForLocation(locationId, user)
-          .listen((climbs) {
-        add(ClimbsUpdatedEvent(
-            climbs..sort((a, b) => a.section.compareTo(b.section))));
+      climbSubscription =
+          LocationRepo.getInstance().getClimbsForLocation(locationId, user).listen((climbs) {
+            add(ClimbsUpdatedEvent(climbs..sort((a, b) => a.section.compareTo(b.section))));
       });
     } else {
       add(LocationErrorEvent());
@@ -37,9 +37,8 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
     add(ClearSectionsEvent());
     final user = await UserRepo.getInstance().getCurrentUser();
     if (user != null) {
-      locationSubscription = LocationRepo.getInstance()
-          .getSectionsForLocation(locationId, user)
-          .listen((location) {
+      locationSubscription =
+          LocationRepo.getInstance().getSectionsForLocation(locationId, user).listen((location) {
         add(SectionsUpdatedEvent(location.sections));
       });
     } else {
@@ -54,23 +53,25 @@ class LocationBloc extends Bloc<LocationEvent, LocationState> {
     return LocationState.initial();
   }
 
+  void retrieveClimb(Climb climb, LocationWidget view) async {
+    final currentUser = await UserRepo.getInstance().getCurrentUser();
+    ClimbRepo.getInstance().getClimb(climb, currentUser).then((climb) {
+      view.navigateToClimb(climb);
+    });
+  }
+
   @override
   Stream<LocationState> mapEventToState(LocationEvent event) async* {
     if (event is ClearClimbsEvent) {
-      yield LocationState.isLoading(
-          true, <Climb>[], state.sections, state.grades);
+      yield LocationState.isLoading(true, <Climb>[], state.sections, state.grades);
     } else if (event is ClimbsUpdatedEvent) {
-      yield LocationState.isLoading(
-          false, event.climbs, state.sections, state.grades);
+      yield LocationState.isLoading(false, event.climbs, state.sections, state.grades);
     } else if (event is ClearSectionsEvent) {
-      yield LocationState.isLoading(
-          true, state.climbs, <String>[], state.grades);
+      yield LocationState.isLoading(true, state.climbs, <String>[], state.grades);
     } else if (event is SectionsUpdatedEvent) {
-      yield LocationState.isLoading(
-          false, state.climbs, event.sections, state.grades);
+      yield LocationState.isLoading(false, state.climbs, event.sections, state.grades);
     } else if (event is LocationErrorEvent) {
-      yield LocationState.isLoading(
-          false, state.climbs, state.sections, state.grades);
+      yield LocationState.isLoading(false, state.climbs, state.sections, state.grades);
     }
   }