+++ lib/models/attempt.dart
@@ -1,8 +1,11 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+
 class Attempt {
-  Attempt(this.timestamp, this.sendType, this.warmup,
+  Attempt(this.id, this.timestamp, this.sendType, this.warmup,
       this.drills, this.notes);
 
-  final DateTime timestamp;
+  final String id;
+  final Timestamp timestamp;
   final String sendType;
   final bool warmup;
   final List<String> drills;