+++ lib/models/location_repo.dart
@@ -26,8 +26,7 @@ class LocationRepo {
   Stream<List<Location>> getLocationsForUser(User user) {
     return _firestore
         .collection(
-        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths
-            .LOCATIONS_SUBPATH}")
+        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths.LOCATIONS_SUBPATH}")
         .snapshots()
         .map((data) => Deserializer.deserializeLocations(data.documents));
   }
@@ -35,8 +34,7 @@ class LocationRepo {
   Future<SelectedLocation> getLocation(Location location, User user) async {
     DocumentReference locationRef = _firestore
         .collection(
-        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths
-            .LOCATIONS_SUBPATH}")
+        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths.LOCATIONS_SUBPATH}")
         .document(location.id);
     if (locationRef != null) {
       try {
@@ -53,8 +51,7 @@ class LocationRepo {
   Stream<List<Climb>> getClimbsForLocation(String locationId, User user) {
     return _firestore
         .collection(
-        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths
-            .CLIMBS_SUBPATH}")
+        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths.CLIMBS_SUBPATH}")
         .where("locationId", isEqualTo: locationId)
         .snapshots()
         .map((data) => Deserializer.deserializeClimbs(data.documents));
@@ -63,8 +60,7 @@ class LocationRepo {
   Stream<Location> getSectionsForLocation(String locationId, User user) {
     return _firestore
         .collection(
-        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths
-            .LOCATIONS_SUBPATH}")
+        "${FirestorePaths.USERS_COLLECTION}/${user.uid}/${FirestorePaths.LOCATIONS_SUBPATH}")
         .document(locationId)
         .snapshots()
         .map((data) {